perbaikan tombol hapus data keluarga mengikuti aturan data lengkap (#3073)

+++ donjo-app/controllers/Keluarga.php
@@ -109,6 +109,7 @@ class Keluarga extends Admin_Controller
         }
 
         $data['func']       = 'index';
+        $this->header['kategori']     = 'data_lengkap';
         $data['set_page']   = $this->_set_page;
         $list_data          = $this->keluarga_model->list_data($o, $p);
         $data['paging']     = $list_data['paging'];
@@ -415,6 +416,13 @@ class Keluarga extends Admin_Controller
     public function delete($p = 1, $o = 0, $id = 0)
     {
         $this->redirect_hak_akses('h');
+
+        if (data_lengkap()) {
+            session_error('Data tidak dapat proses karena sudah dinyatakan lengkap');
+
+            redirect("{$this->controller}/index/{$p}/{$o}");
+        }
+
         $this->redirect_tidak_valid($this->keluarga_model->cek_boleh_hapus($id));
         $this->keluarga_model->delete($id);
         $this->cache->hapus_cache_untuk_semua('_wilayah');
@@ -425,6 +433,13 @@ class Keluarga extends Admin_Controller
     public function delete_all()
     {
         $this->redirect_hak_akses('h');
+
+        if (data_lengkap()) {
+            session_error('Data tidak dapat proses karena sudah dinyatakan lengkap');
+
+            redirect("{$this->controller}/index/{$p}/{$o}");
+        }
+        
         $this->keluarga_model->delete_all();
         $this->cache->hapus_cache_untuk_semua('_wilayah');
